services/app-one/requirements.txt
flask==2.2.2
python-dotenv==0.21.0
gunicorn==20.1.0